Port typing tests to bazel test
