+++ python/ray/tests/typing_files/check_typing_bad.py
@@ -18,7 +18,7 @@ def h(a: str, b: int) -> str:
     return a
 
 
-# Does not typecheck:
+# Does not typecheck due to incorrect input type:
 a = h.remote(1, 1)
 b = f.remote("hello")
 c = f.remote(1, 1)
@@ -27,4 +27,7 @@ d = f.remote(1) + 1
 # Check return type
 ref_to_str = f.remote(1)
 unwrapped_str = ray.get(ref_to_str)
-unwrapped_str + 100  # Should fail type check
+unwrapped_str + 100  # Fail
+
+# Check ObjectRef[T] as args
+f.remote(ref_to_str)  # Fail
